+++ src/camera.h
@@ -15,11 +15,11 @@ class Camera {
 
 class PerspectiveCamera : public Camera {
     public:
-    PerspectiveCamera(const Transform *CameraToWorld, const Point2i &resolution, float fov = 60); // we add fov later
+    PerspectiveCamera(const Transform *CameraToWorld, const Point2i &resolution, double fov = 60); // we add fov later
     virtual Ray GenerateRay(const Point2f &sample, const Point2i &pixel) const;
     
     private:
-    float fov;
+    double fov;
     Point2f lowerLeft;
 };
