+++ src/geometry/matrix.cpp
@@ -6,17 +6,17 @@
 
 
 Matrix4x4::Matrix4x4(){
-    m[0][0] = m[1][1] = m[2][2] = m[3][3] = 1.f;
+    m[0][0] = m[1][1] = m[2][2] = m[3][3] = 1;
     m[0][1] = m[0][2] = m[0][3] = m[1][0] = m[1][2] = m[1][3] = m[2][0] =
-            m[2][1] = m[2][3] = m[3][0] = m[3][1] = m[3][2] = 0.f;
+            m[2][1] = m[2][3] = m[3][0] = m[3][1] = m[3][2] = 0;
 }
 
-Matrix4x4::Matrix4x4(float mat[4][4]) { memcpy(m, mat, 16 * sizeof(float)); }
+Matrix4x4::Matrix4x4(double mat[4][4]) { memcpy(m, mat, 16 * sizeof(double)); }
 
-Matrix4x4::Matrix4x4(float t00,float t01,float t02,float t03,
-         float t10,float t11,float t12,float t13,
-         float t20,float t21,float t22,float t23,
-         float t30,float t31,float t32,float t33){
+Matrix4x4::Matrix4x4(double t00,double t01,double t02,double t03,
+         double t10,double t11,double t12,double t13,
+         double t20,double t21,double t22,double t23,
+         double t30,double t31,double t32,double t33){
     m[0][0] = t00;
     m[0][1] = t01;
     m[0][2] = t02;
@@ -81,18 +81,18 @@ Matrix4x4 Transpose(const Matrix4x4 &m){
 Matrix4x4 Inv(const Matrix4x4 &m){
     int indxc[4], indxr[4];
     int ipiv[4] = {0, 0, 0, 0};
-    float minv[4][4];
-    memcpy(minv, m.m, 4 * 4 * sizeof(float));
+    double minv[4][4];
+    memcpy(minv, m.m, 4 * 4 * sizeof(double));
     for (int i = 0; i < 4; i++) {
         int irow = 0, icol = 0;
-        float big = 0.f;
+        double big = 0;
         // Choose pivot
         for (int j = 0; j < 4; j++) {
             if (ipiv[j] != 1) {
                 for (int k = 0; k < 4; k++) {
                     if (ipiv[k] == 0) {
                         if (std::abs(minv[j][k]) >= big) {
-                            big = float(std::abs(minv[j][k]));
+                            big = double(std::abs(minv[j][k]));
                             irow = j;
                             icol = k;
                         }
@@ -108,17 +108,17 @@ Matrix4x4 Inv(const Matrix4x4 &m){
         }
         indxr[i] = irow;
         indxc[i] = icol;
-        if (minv[icol][icol] == 0.f) Error("Singular matrix in MatrixInvert");
+        if (minv[icol][icol] == 0) Error("Singular matrix in MatrixInvert");
 
         // Set $m[icol][icol]$ to one by scaling row _icol_ appropriately
-        float pivinv = 1. / minv[icol][icol];
+        double pivinv = 1. / minv[icol][icol];
         minv[icol][icol] = 1.;
         for (int j = 0; j < 4; j++) minv[icol][j] *= pivinv;
 
         // Subtract this row from others to zero out their columns
         for (int j = 0; j < 4; j++) {
             if (j != icol) {
-                float save = minv[j][icol];
+                double save = minv[j][icol];
                 minv[j][icol] = 0;
                 for (int k = 0; k < 4; k++) minv[j][k] -= minv[icol][k] * save;
             }
