+++ src/geometry/matrix.h
@@ -3,13 +3,13 @@
 
 
 struct Matrix4x4{
-    float m[4][4];
+    double m[4][4];
     Matrix4x4();
-    Matrix4x4(float mat[4][4]);
-    Matrix4x4(float t00,float t01,float t02,float t03,
-         float t10,float t11,float t12,float t13,
-         float t20,float t21,float t22,float t23,
-         float t30,float t31,float t32,float t33);
+    Matrix4x4(double mat[4][4]);
+    Matrix4x4(double t00,double t01,double t02,double t03,
+         double t10,double t11,double t12,double t13,
+         double t20,double t21,double t22,double t23,
+         double t30,double t31,double t32,double t33);
 
     Matrix4x4 operator*(const Matrix4x4 &m2) const;
