+++ src/shape.h
@@ -9,7 +9,7 @@
 class Shape {
     public:
     Shape(const Transform *ObjectToWorld, const Transform *WorldToObject) : ObjectToWorld{ObjectToWorld}, WorldToObject{WorldToObject} {};
-    virtual bool Intersect(const Ray &r, float *tHit, SurfaceInteraction *isect) const = 0;
+    virtual bool Intersect(const Ray &r, double *tHit, SurfaceInteraction *isect) const = 0;
     virtual ~Shape(){};
     // virtual Bounds3f ObjectBounds() const = 0;
     // virtual Bounds3f WorldBounds() const;
